==> r-code/02_Whiteplague_workingdata.Rmd <==
---
title: "Creation of white_working_data"
author: "Andia Chaves Fonnegra"
date: "7/18/2017"
output: pdf_document
---


```{r setup, include=FALSE, warnings=FALSE}
library(knitr)
opts_knit$set(root.dir='../')                     # definining working directory; or normalizePath('../')
opts_chunk$set(fig.align='center',                # aligns all figures
                echo=TRUE,                        # shows r-code
                message=FALSE,                    # suppresses library outputs
                warnings=FALSE,                   # suppresses library outputs
                tidy=TRUE,  # prevents the source code from running off a pdf page
                dev='pdf')                        # pdf device
```

# Introduction
This file is to prepare the data before feeding it into the model.
Data was scaled and saved as working_data file to feed into the model


## Loading data
```{r}
rm(list=ls(all=TRUE))
detachAllPackages() 
# load data & libraries - Loading Raw data file as r.data
load(file="data/Whiteplague_rawdata.Rdata")
ls()
#check data
dim(Whiteplague_rawdata)
head(Whiteplague_rawdata)
library(lubridate)
```


## Data preparation
```{r data_preparation}
#All environmental data is scaled due to difference in units, this to avoid adding more variability in the analysis
Whiteplague_rawdata$Chl_sca        <- scale(Whiteplague_rawdata$Chlorophyll)
Whiteplague_rawdata$Cond_sca       <- scale(Whiteplague_rawdata$Conductivity)
Whiteplague_rawdata$Den_sca        <- scale(Whiteplague_rawdata$Density)
Whiteplague_rawdata$DepCTD_sca     <- scale(Whiteplague_rawdata$Depth_CTD)
Whiteplague_rawdata$Oxy_sca        <- scale(Whiteplague_rawdata$Oxygen)
Whiteplague_rawdata$Sal_sca        <- scale(Whiteplague_rawdata$Salinity)
Whiteplague_rawdata$Turb_sca       <- scale(Whiteplague_rawdata$Turbidity)
Whiteplague_rawdata$Depth_tran_sca <- scale(Whiteplague_rawdata$Depth_transect)
Whiteplague_rawdata$Temp_sca       <- scale(Whiteplague_rawdata$Temperature)
Whiteplague_rawdata$DHW_sca        <- scale(Whiteplague_rawdata$Max_DHW)
Whiteplague_rawdata$Rain_sca       <- scale(Whiteplague_rawdata$Rainfall)
Whiteplague_rawdata$Depth          <- Whiteplague_rawdata$Depth_transect
```


## Time point

```{r}
#creating a time lag between data points to be able to add temporal autocorrelation in the model

# #first switch formate to date formt for Time_point
Whiteplague_rawdata$Time_point <- as.Date(Whiteplague_rawdata$Time_point,format="%d_%b-%y")
# Whiteplague_rawdata$Time_point
# #Then create a new column with the  time lag (i-1)
# Whiteplague_rawdata["Time_lag"] <-NA
# # create a function that calculates the lag for n periods
# lag <- function(x, n) c(rep(NA, n), x[1:(length(x) - n)])
# # add data to the new column Time_lag
# Whiteplague_rawdata$Time_lag <- as.numeric(Whiteplague_rawdata$Time_point- lag(Whiteplague_rawdata$Time_point, 1))


# in Zuur: (seq(0,1000000000,length=365))
Whiteplague_rawdata$TimePoint <- as.numeric(paste(Whiteplague_rawdata$Year,yday(Whiteplague_rawdata$Time_point),sep="."))
```


```{r}

#Sorting the data frame so Months can be in numerical order

Whiteplague_workingdata<-Whiteplague_rawdata[order(Whiteplague_rawdata$Month_12),]
Whiteplague_workingdata

ls()
#check data
ls(Whiteplague_workingdata)
dim(Whiteplague_workingdata)
head(Whiteplague_workingdata)

#count how many samples per month we have in total this is to later define in the model
length(table(Whiteplague_workingdata$Month_12))
```



#aggregate working_data file per year
```{r}
#Make a new data frame only with annual mean values for each variables, in this way we can join biological variables (that only were collected once a year) with environmental variables (here as mean per season, per year).
library(dplyr)
#packageVersion("dplyr")

#Group data per station, year and season and summarize the max value of each variable
a1<- Whiteplague_workingdata %>% 
  group_by(Station_code, Year, Season,Latitude, Longitude) %>% 
  summarise (Coral = max(Coral) , Sand = max(Sand), Sponges = max(Sponges), Macroalgae = max(Macroalgae), Cyanobacteria = max(Cyanobacteria),ill = max(ill), Ntot=max(Ntot),Plague = max(Plague),Chl_sca = max(Chl_sca),Oxy_sca = max(Oxy_sca),Sal_sca = max(Sal_sca),Turb_sca = max(Turb_sca),Depth_tran_sca = max(Depth_tran_sca),Temp_sca = max(Temp_sca), DHW_sca = max(DHW_sca) )
head(a1)


# subset data to keep only the max value for season 4 (rainny-heat), mantaining all the stations and years and previous data

Whiteplague_workingdata_Annual_orig<- filter(a1, Season %in% 4)
df <- Whiteplague_workingdata_Annual_orig
df$Sand <- scale(df$Sand)
df$Coral <- scale(df$Coral)
df$Sponges <- scale(df$Sponges)
df$Macroalgae <- scale(df$Macroalgae)
df$Cyanobacteria <- scale(df$Cyanobacteria)

Whiteplague_workingdata_Annual <- df

#Finally!!
```



## Saving data
```{r}
#Save working data  file with data already scaled/ and or transformed to be used by the model
save(
Whiteplague_workingdata,
Whiteplague_workingdata_Annual_orig,Whiteplague_workingdata_Annual,
file="data/Whiteplague_workingdata.Rdata")
```
