==> r-code/02.2_Figure 2. Boxplot and Kruskal Wallis test.Rmd <==
---
title: "Figure 2. Boxplot, linear graphs and also tet of normality and ANOVA"
author: "Andia Chaves Fonnegra"
date: "2/21/2018"
output: pdf_document
---


```{r setup, include=FALSE, warnings=FALSE}
library(knitr)
opts_knit$set(root.dir='../')                     # definining working directory; or normalizePath('../')
opts_chunk$set(fig.align='center',                # aligns all figures
                echo=TRUE,                        # shows r-code
                message=FALSE,                    # suppresses library outputs
                warnings=FALSE,                   # suppresses library outputs
                tidy=TRUE,  # prevents the source code from running off a pdf page
                dev='pdf')                        # pdf device
```

#Introduction

This document provides the code for Box Plot Figure 2 of the paper.

#Data
The white plague prevalence data and corresponding environmental and biological variables used in this model can be downloaded from xxxxxxx. The file "whiteplague_working.RData" was used for this boxplot. Not the rawdata file.

## Load data
```{r}
rm(list=ls(all=TRUE))
# Load settings
source("r-code/00_settings.r")

load(file="data/Whiteplague_workingdata.Rdata")
head(Whiteplague_workingdata)
dim(Whiteplague_workingdata)
```

#Normality test

```{r}
#Shapiro-Wilkinson normality test
shapiro.test(Whiteplague_workingdata$Plague)
#for producing a normal quantile-quantile plot
qqnorm(Whiteplague_workingdata$Plague)
qqline(Whiteplague_workingdata$Plague)

#If p> than 0.05. Data is normal, if not, not normal
```
Prevalence data of disease is no-normal distribute- go with Kruskal Wallis test.

#Bartlest test of homegenty of variances

```{r}
bartlett.test(Whiteplague_workingdata$Plague~Whiteplague_workingdata$Station_code)
#If p> than 0.05. variances are homogeneous, if not, not homogeneous
```
WP prevalence of disease data is non-homogeneous


#If normal One way ANOVA comparing between the habitats

```{r}
summary(aov(area ~ time, data=grow)) #correct data depending of file to use
#To do the Tukey test a posteriori
TukeyHSD(results, conf.level=0.95)
#This is to make the regular boxplot with the two variables the $ connects the file name with the column within the file
boxplot(grow$area ~ grow$time)
qqnorm(Reefvariables$Coral.Density)
```


#if no-normal kruskal wallis

```{r}
#To do kruskal Wallis test, pecify each column withn the file so it can be read as a vector and not as data frame
kruskal.test(Whiteplague_workingdata$Plague~Whiteplague_workingdata$Station_code)

library(dunn.test)
#To do Pots-hoc dunn and Kruskal Wallis in one
dunn.testresults<-dunn.test (Whiteplague_workingdata$Plague, g=Whiteplague_workingdata$Station_code, method="bonferroni", kw=TRUE, label=TRUE, wrap=TRUE, rmc=FALSE, alpha=0.05)


boxplot(Whiteplague_workingdata$Plague ~ Whiteplague_workingdata$Station_code, las=2)
text(3,1,"*",cex=0.5)

```

#FIGURE 2 MANUSCRIPT 

#FIRST Add a column of depth category and month category to make easier the figure (graph)

```{r}

#Add depth category
Whiteplague_workingdata$Depth_categories <- cut(Whiteplague_workingdata$Depth_transect, breaks = c(0,15,30,45), labels = c("Shallow","Mid-depth", "Mesophotic"), right = FALSE)

Whiteplague_workingdata$Station_code
Whiteplague_workingdata$Depth_transect [0:100]
Depth_categories [0:100]

# Month
Whiteplague_workingdata$Month <- cut(Whiteplague_workingdata$Month_12, breaks = c(0,1.1,2.1,3.1,4.1,5.1,6.1,7.1,8.1,9.1,10.1,11.1,12.1), labels = c("J","F","Mr","A","My","Jn","Jl","A","S","O","N","D"), right = FALSE, left=FALSE)

Whiteplague_workingdata$Month_12 [0:100]
Whiteplague_workingdata$Month [0:100]

```

#Fig2. A) WP prevalence per station per depth two options

#Option 1, is use depth categories on x axis. Although I would prefer option 2) stations on x axis (However, option two do not organize the stations per depth, even that I rearange their order)

```{r}
#Arranging order of stations to match depth (reorder only arrange by median value, not depth value)
library(forcats)
library(ggthemes)
library(ggplot2)
library(devtools)


theme_set(theme_classic())
#Option 1
ggplot(Whiteplague_workingdata, aes(x = reorder(Depth_categories, Plague, fun = median), y = Plague,fill=Station_code)) + geom_boxplot()+
theme(axis.text.x = element_text(angle=360, vjust=0.8, size=16), axis.text.y = element_text(angle=360, vjust=0.4, size=14), legend.key.size = unit(1.1, "cm"), axis.title.y= element_text(size=16,face="bold", margin = margin(t = 20, r = 30, b = 0, l = 10)), axis.title.x= element_text(size=16,face="bold", margin = margin(t = 30, r = 20, b = 10, l = 0))) + labs(x= "Stations", y="White Plague Prevalence")

#Option 2) I prefer this graph, however I would like that moddepth stations will be together in teh middle, and not in different positions. ahh.

ggplot(Whiteplague_workingdata, aes(x = reorder(Station_code, Plague, fun = median), y = Plague,fill=Depth_categories)) + geom_boxplot()+
theme(axis.text.x = element_text(angle=360, vjust=0.6)) + 
  labs(x= "Stations", y="White Plague Prevalence")
```

# Fig 2. B) Prevalence of disease per month per depth category

```{r}
#Horrible, discusting graph, seems to need some codding to diferentiate standard error from average.

WPplot <- data.frame(Whiteplague_workingdata %>% group_by(Depth_categories,Time_point) %>% summarise(WPmean=mean(Plague),WPsd=sd(Plague)))
pd <- position_dodge(5) # move them .05 to the left and right
ggplot(WPplot, 
       aes(x=Time_point, y=WPmean, colour=Depth_categories, group=Depth_categories)) +
  geom_errorbar(aes(ymin=pmax(WPmean-WPsd,0), ymax=WPmean+WPsd), colour="black", width=50, position=pd) +
  geom_line() +
  geom_point(position=pd, size=3, shape=21, fill="white") + # 21 is filled circle
  xlab("Time point") +
  ylab("White Plague prevalence") +
scale_colour_hue(name="Sea depth",    # Legend label, use darker colors
                     breaks=c("Shallow","Mid-depth", "Mesophotic"),
                     labels=c("Shallow", "Mid-depth", "Mesophotic"),
                     l=40) +                    # Use darker colors, lightness=40
    ggtitle("Average monthly pattern of White Plague disease prevalence for different sea depth") +
    #scale_y_continuous(limits = c(0, NA))+
    theme_bw() +
    theme(legend.justification=c(1,0),legend.position=c(1,0), # Position legend in bottom right
          plot.title = element_text(hjust = 0.5)) # title position - justified
```

# Fig 2. C) Prevalence of disease per season per depth category


```{r}
#Same problem with season, no clear average lines and standard error, check on codding. Compare with excel graphs.
ggplot(Whiteplague_workingdata, 
       aes(x=Season, y=Plague, group=Depth_categories)) + geom_line((aes(color=Depth_categories)))

```
