==> r-code/03_Whiteplague_rstan_bio_Model.Rmd <==
---
title: "Environmental drivers of white plague disease on shallow and mesophotic coral reefs, U.S. Virgin Islands- Annual Model Biological and Environmental combined"
author: 
-name:"Andia Chaves Fonnegra and Bernd Panassiti"
  affiliation: University of the Virgin Islands (US.VI) and Laimburg Research Center (Italy)
  abstract: "This document provides the r-code for the white plague disease model described on the article:"Environmental drivers of white plague disease on shallow and mesophotic coral reefs, U.S. Virgin Islands" Description of the model in terms of data collection, white plague disease, parameter estimations and fitting of the model can be found in Appendix A."
  keywords: white plague, coral reefs, bayesian model, environment
date: "7/18/2017"
output: pdf_document
number_sections: yes
    toc: yes
header-includes: \usepackage{graphicx}
email: \email{}
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE, warnings=FALSE}
library(knitr)
opts_knit$set(root.dir='../')                     # definining working directory; or normalizePath('../')
opts_chunk$set(fig.align='center',                # aligns all figures
                echo=TRUE,                        # shows r-code
                message=FALSE,                    # suppresses library outputs
                warnings=FALSE,                   # suppresses library outputs
                tidy=TRUE,  # prevents the source code from running off a pdf page
                dev='pdf')                        # pdf device
```

#Introduction

This document provides the r-code for the white plague disease model described on the article.The model address to the objectives of this study: (1) to evaluate the annual patterns of WP disease, and (2) to identify and quantify the influence of environmental and BIOLOGICAL factors on WP disease prevalence using Bayesian inference. We use rstan for statistical inference (http://mc-stan.org/about/). 
The model uses abiotic (without quadratic term) and biotics and year as random effect.

#Data
The white plague prevalence data and corresponding environmental and biological variables used in this model can be downloaded from xxxxxxx. The file "whiteplague_working.RData" needs to be placed in a folder called "Data".
"

#Model

## Load data
```{r}
rm(list=ls(all=TRUE))
# Load settings
source("r-code/00_settings.r")

load(file="data/Whiteplague_workingdata.Rdata")

head(Whiteplague_workingdata)
```


##Load libraries and create directory for results
```{r}
# Date format - forms part of names of created files or graphs
today <- format(Sys.time(), "%Y.%m.%d")
set.seed(2017)

# load libraries
library(formatR)
library(StanHeaders)
library(rstan)
rstan_options(auto_write = TRUE) # automatically save the compilied Stan program
options(mc.cores = parallel::detectCores()) # execute multiple Markov chains in paral

library(MCMCpack) # rwish function
library(modeest)
library(IDPmisc)
library(ggplot2) # for graphs
library(DHARMa) #to extract posterior
```



## Data preparation
```{r}
df <- data.frame(Whiteplague_workingdata_Annual)
df <- df[,c("Station_code","Year","Season","Depth_tran_sca", "Ntot","ill","Plague","Chl_sca","Oxy_sca","Sal_sca","Turb_sca","Temp_sca","DHW_sca", "Sand", "Coral", "Sponges","Macroalgae", "Cyanobacteria")]

df$healthy  <- df$Ntot-df$ill
df$Year     <- as.factor(Whiteplague_workingdata_Annual$Year)
df$Site     <- as.factor(as.numeric(Whiteplague_workingdata_Annual$Station_code))
df$Obsid    <- as.factor(1:nrow(df))


df$Site_index  = as.integer(df$Site)
df$Year_index  = as.integer(df$Year)
df$Depth_index = as.integer(df$Depth)
df$Obsid_index = as.integer(df$Obsid)



Environment <- df[,c("Chl_sca","Oxy_sca","Sal_sca","Turb_sca","Temp_sca", "DHW_sca","Depth_tran_sca","Sand","Coral","Sponges","Macroalgae", "Cyanobacteria")]
```


## Model definition
```{r}
dataList = list(
  TotalObservedcolonies = df$Ntot, #Total observed colonies Shallow (2 years) Deep (4 years)
  Diseasedcolonies = df$ill, # Total numer of disease colonies Shallow (2 years) Deep (4 years)
  nWP = nrow(df), #White Plague prevalence total number of rows
  observedWpPrevalence = df$Plague, # Total Prevalence of White Plague Shallow (2 years) Deep (4 years)
  
  Environment = Environment,
  nEnvi = ncol(Environment),
  
  Year = df$Year_index,
  nYears =length(unique(df$Year_index)),
  
  Site = df$Site_index,
  nSites =length(unique(df$Site_index))
       
)
```


## Running the model

```{r model_run,eval=FALSE}
time <- proc.time()
set.seed(2021)
stan.white <- stan("r-code/03_Whiteplague_rstan_bio_Model.stan",
                   data = dataList,
                   iter = 30000, 
                   chains = 4,
                   control = list(adapt_delta = 0.99,max_treedepth = 15),
                   refresh=10)
(proc.time() - time)/60/60
```


#Save the model run results
```{r}
save(stan.white, file ="results/2021.01.18_Whiteplague_Biological_model_result.Rdata")
```


```{r}
fit = stan.white
names(fit)[2:13] = names(Environment)
plot(fit,pars=names(Environment)) # with environment

plot(fit,pars=names(Environment)[8:12])

summary(fit,pars=names(Environment))$summary
```




```{r}
posterior<-extract(stan.white,inc_warmup=F,permute=F)

Ntot <- Whiteplague_workingdata_Annual$Ntot
observedWhitePlague <- Whiteplague_workingdata_Annual$ill

Nchains=4 # number of chains, see model definition
modelpredictionsChains <- paste(rep("modelPredictions[,,",Nchains),1:Nchains,rep("]",Nchains),sep="")


# FITTED PREDICTED
modelPredictions <- NULL
modelPredictions<-apply(posterior, MARGIN=1:2, FUN = function(draw) draw[grepl("^RealDisease[\\[]",names(draw))])


# combine 4 chains
predictionsLinearWhitePlague <- NULL
assign("predictionsLinearWhitePlague",eval(parse(text=paste("cbind(",paste(modelpredictionsChains, collapse = ","),")",sep=""))))

predictionsResponseMedianWhitePlague <- inv.logit(as.vector(apply(predictionsLinearWhitePlague, 1, median)))*Ntot

# SIMULATED
modelPredictions <- NULL
modelPredictions<-apply(posterior, MARGIN=1:2, FUN = function(draw) draw[grepl("^DiseasePredictions[\\[]",names(draw))])

# combine 4 chains
simulatedPredictedWhitePlague <- NULL
assign("simulatedPredictedWhitePlague",eval(parse(text=paste("cbind(",paste(modelpredictionsChains, collapse = ","),")",sep=""))))


WhitePlagueDHARMaResponse <- createDHARMa(simulatedResponse = simulatedPredictedWhitePlague,
  observedResponse = observedWhitePlague,fittedPredictedResponse = predictionsResponseMedianWhitePlague,integerResponse = T)


plot(WhitePlagueDHARMaResponse)
```


```{r}
tableSI =data.frame(sites=Station_code,WhitePlagueDHARMaResponse$observedResponse,WhitePlagueDHARMaResponse$fittedPredictedResponse,WhitePlagueDHARMaResponse$scaledResiduals)
write.table(tableSI,"results/2021.01.18-BiologicalModel_tableSI.csv", sep=",", col.names=TRUE)
```
