==> r-code/04.1_Whiteplague_rstanModel_Extraction_RunFirst.Rmd <==
---
title: "Extraction from posterior"
author: "Bernd Panassiti"
date: 'created: 26.07.2017, last modified: `r format(Sys.Date(), format="%d.%m.%Y")`'
output: html_document
header-includes: \usepackage{graphicx}
---

## Introduction
Script extracts necessary data to create diagnositic plots for the RStan model for White Plague disease.
Needs as input the stan.white object.  


## Initial setup
```{r setup, include=FALSE, warnings=FALSE}
library(knitr)
opts_knit$set(root.dir='../')       # definining working directory; or normalizePath('../')
opts_chunk$set(fig.align='center',   # aligns all figures
                echo=FALSE,          # suppresses r-code
                message=FALSE,       # suppresses library outputs
                warnings=FALSE,      # suppresses library outputs
                dev='pdf')           # pdf device
```

```{r load}
rm(list=ls(all=TRUE))

set.seed(123)

# load settings, data & libraries
source("r-code/00_settings.r")

# unload all packages in "other attached packages"
sessionInfo()
#detach("package:",unload=T)

library(plyr) # aplyr (function:stanGraphInput)
library(rstan) # extraction of posterior, function:stanGraphInput
library(MCMCpack) # rwish function, to create mmcmc list, function:stanGraphInput
library(zoo) # rollapply


#Load files workingdata and stan file with the results of the model
load(file="data/Whiteplague_workingdata.Rdata")

NameOfModelRun <- "2021.01.17_Whiteplague_Biological_model_result" # last biological
#NameOfModelRun <- "2018.03.08_Whiteplague_Environmental_model_result" # last environmental
load(file=paste("results/",NameOfModelRun,".Rdata",sep=""))

```


## Data preparation for graphs

### Define Observed data

```{r ObsPred,,echo=FALSE}
observedWhitePlague<-Whiteplague_workingdata_Annual$Plague
Ntot <- Whiteplague_workingdata_Annual$Ntot
# observedWhitePlague<-Whiteplague_workingdata$Plague
# Ntot <- Whiteplague_workingdata$Ntot
```


### Extraction of coefficients from stan object and of posterior
```{r WhitePlague_graph_input}
GraphInput<-stanGraphInput(stan.white) # reguires libraries: rstan, MCMCpack, plyr
dnam <- GraphInput$parameters
# # extract entire posterior; all parameters
posterior<-extract(stan.white,inc_warmup=F,permute=F)
```




# Predictions
## Posterior predictive simulations - Simulated Response
```{r Predictions_posterior_simulatedResponse}
# # posterior is nIter (e.g. 1000) iterations by nChain (e.g. 4) chains by nPar (e.g. 20) parameters. 
# # So, grepl("^realVectorPredictions",dimnames(posterior)[[3]])] extracts the parameters whose first two letters are realVectorPredictions, 
# # leaving a nIter by nChain by nPar array. The apply() command applies the specified function to the specified dimension of the resulting array, 
# # in this case the mean function to the third dimension
# dimnames(posterior) # lists all objects
# get median of predictions,e.g.: as.vector(apply(predictedPathogen, 1, median))
modelPredictions <- NULL
modelPredictions<-apply(posterior, MARGIN=1:2, FUN = function(draw) draw[grepl("^DiseasePredictions[\\[]",names(draw))])
# combine 4 chains
simulatedPredictedWhitePlague <- NULL
simulatedPredictedWhitePlague <- cbind(modelPredictions[,,1], modelPredictions[,,2], modelPredictions[,,3], modelPredictions[,,4]) #  combining chains, see also: combine.mcmc from runjags package
```


## Median model predictions on the linear / response scale - Fitted predicted response
```{r Predictions_posterior_fittedPredictedResponse}
# White Plague Prevalence
modelPredictions <- NULL
modelPredictions<-apply(posterior, MARGIN=1:2, FUN = function(draw) draw[grepl("^RealDisease[\\[]",names(draw))])
# combine 4 chains
predictionsLinearWhitePlague <- NULL
predictionsLinearWhitePlague <- cbind(modelPredictions[,,1], modelPredictions[,,2], modelPredictions[,,3], modelPredictions[,,4]) #  combining chains, see also: combine.mcmc from runjags package
predictionsResponseMedianWhitePlague <- inv.logit(as.vector(apply(predictionsLinearWhitePlague, 1, median)))*Ntot
```


### Extraction of coefficients
```{r graph_coefficients}
# define which rows to extract from GraphInput
modelParameterLength <- length(which(grepl("b0|par",dnam))) # total number of parameters from all models
#Vector
Nrows<-c(1:length(rep(seq(1:modelParameterLength),each=4)))
# Medians
varMedianPlotWhitePlague <- GraphInput$var.median[Nrows,]
datqPlotWhitePlague <- GraphInput$datq
```


##Save files
```{r save}
save(
posterior,GraphInput,
# Observations
observedWhitePlague,
# Predictions
# 1. Posterior predictive simulations
simulatedPredictedWhitePlague,
# 2. Median model predictions on the linear / response scale
predictionsLinearWhitePlague,
predictionsResponseMedianWhitePlague,
# Parameter names
#dnam,
#ParameterNamesWhitePlague,
# Coefficients
#varMedianPlotWhitePlaque,datqPlotWhitePlaque,
varMedianPlotWhitePlague,datqPlotWhitePlague,
file=paste("results/",NameOfModelRun,"_graphinput.Rdata",sep=""))
```
