==> r-code/07_Appendix.Rmd <==
---
title: Elevated seawater turbidity and temperature as drivers of white plague disease
  on shallow and mesophotic coral reefs - Appendix A
author: 
  - Andia Chaves-Fonnegra
  - Bernd Panassiti
  - Tyler B. Smith
  - Elizabeth Brown
  - Elizabeth Clemens
  - Moriah Sevier
  - Marilyn Brandt
keywords: white plague, coral decline, coral reefs, Bayesian inference
date: 'created: 7/18/2017, last modified: `r format(Sys.Date(), format="%m/%d/%Y")`'
email: \email{andiachaves@gmail.com}
output: pdf_document
---

```{r setup, include=FALSE, warnings=FALSE}
library(knitr)
opts_knit$set(root.dir='../')                     # definining working directory; or normalizePath('../')
opts_chunk$set(fig.align='center',                # aligns all figures
                echo=FALSE,                       # suppresses r-code
                message=FALSE,                    # suppresses library outputs
                warnings=FALSE,                   # suppresses library outputs
                tidy=TRUE,  # prevents the source code from running off a pdf page
                dev='pdf')                        # pdf device
```

#Introduction
This appendix contains additional information to the article ``Elevated seawater turbidity and temperature as drivers of white plague disease on shallow and mesophotic coral reefs''.  


This document provides the R-code for the ``Environmental model'', a statistical model that relates abiotic environmental variables to the white plague prevalences on corales in the Caribbean. The objectives of our study were to: (1) to evaluate the annual patterns of WP disease, and (2) to identify and quantify the influence of environmental factors on WP disease prevalence using Bayesian inference. We use RStan for statistical inference (http://mc-stan.org/about/). 

#Data
The white plague prevalence data and corresponding environmental variables used in this model can be downloaded from xxxxxxx. Please note, here we present as an example the model code for the ``Environmental model''.  

However, the data for the ``Biological model'' as described in the article is also included. The code for the model is the same, except that it takes only biological variables as explanatory variables (which were measured annualy) and year and site as random effects.

# White plaque Model
```{r}
# Date format - forms part of names of created files or graphs
today <- format(Sys.time(), "%Y.%m.%d")
set.seed(2017)

# load libraries
library(formatR)
library(StanHeaders)
library(rstan)
rstan_options(auto_write = TRUE) # automatically save the compilied Stan program
options(mc.cores = parallel::detectCores()) # execute multiple Markov chains in paral

# Load settings
source("r-code/00_settings.r")

```

## Defintion of input data
```{r, echo=TRUE, eval=FALSE}
load(file="Whiteplague_workingdata.Rdata")
df <- Whiteplague_workingdata[,c("Ntot","ill","Plague","Chl_sca","Oxy_sca","Sal_sca","Turb_sca","Depth_tran_sca","Temp_sca","DHW_sca")]

df$healthy <- df$Ntot-df$ill
df$Month   <- as.factor(Whiteplague_workingdata$Month_12)
df$Site    <- as.factor(as.numeric(Whiteplague_workingdata$Station_name))
df$Obsid   <- as.factor(1:nrow(df))
df$SiteMonth <- as.factor(as.numeric(as.factor(paste(df$Month,df$Site,sep="-"))))

df$Site_index <- as.integer(df$Site)
df$Month_index=as.integer(df$Month)
df$Obsid_index=as.integer(df$Obsid)
df$SiteMonth_index=as.integer(as.numeric(as.factor(paste(df$Month,df$Site,sep="-"))))

Environment <- df[,c("Chl_sca","Oxy_sca","Sal_sca","Turb_sca","Temp_sca","DHW_sca","Depth_tran_sca")]
```



```{r, echo=TRUE, eval=FALSE}
dataList = list(
  TotalObservedcolonies = df$Ntot, #N obs. colonies Shallow (2 yrs) Deep (4 yrs)
  Diseasedcolonies = df$ill, #N diseased colonies Shallow (2 yrs) Deep (4 yrs)
  nWP = nrow(df), #WP prevalence - N of rows
  observedWpPrevalence = df$Plague, #N Prevalence of WP Shallow (2 yrs) Deep (4 yrs)
  
  Environment = Environment,
  nEnvi = ncol(Environment),
  
  SiteMonth = df$SiteMonth_index,
  nSiteMonth =length(unique(df$SiteMonth_index))
)
```

## Model definition

The following chunk contains the Stan code to run the model. The code needs to be placed in a file called: ``Whiteplague_rstan_env_Model.stan''.

```{r, echo=TRUE, eval=FALSE}
data{
// counters
int<lower=0> nWP;   // number of White Plague cases, int=integer and non negative >0
int<lower=1> nEnvi; // number of environmental variables

int<lower=1> nSiteMonth;               // number of sites-month combinations
int<lower=1,upper=130> SiteMonth[nWP]; // sites-month combinations

matrix[nWP, nEnvi] Environment;

// response
int<lower=0> TotalObservedcolonies [nWP]; // Total number of coral colonies (cases)
int<lower=0> Diseasedcolonies [nWP]; // number of diseased coral colonies (successes)
}

parameters{
real b0; // intercept
// Fixed effects
vector[nEnvi] parEnvironment; // Environmental-level

// Random effects
real<lower=0> sigmalev_1;
vector[nSiteMonth] eta_1;
}

transformed parameters{
vector[nWP] RealDisease;
vector[nSiteMonth] sitemonth_raw;

sitemonth_raw  = eta_1 * sigmalev_1;

RealDisease =  Environment * parEnvironment + sitemonth_raw[SiteMonth];
} //closes transformed parameter

model{
// Fixed effects priors
b0             ~ normal(0, 10);
parEnvironment ~ normal(0, 10);

//Random effects priors
eta_1 ~ normal(0, 100);

sigmalev_1 ~ inv_gamma(0.001, 0.001);

Diseasedcolonies ~ binomial_logit(TotalObservedcolonies,RealDisease);
}

generated quantities{ //predictive inference
real DiseasePredictions[nWP]; //vector to store predictions
for(i in 1:nWP) DiseasePredictions [i] = binomial_rng(TotalObservedcolonies[i],
                                                      inv_logit(RealDisease[i]));
}
```





## Running the model

```{r model_run,eval=FALSE, echo=TRUE}
stan.white <-rstan("Whiteplague_rstan_env_Model.stan",
                      data = dataList,
                      iter = 30000, 
                      chains = 4,
                      control = list(adapt_delta = 0.99,max_treedepth = 15),
                      refresh=10)
```
